note/8_分笔数据实时订阅与多Client问题/v1.0/mac_端/requirements.txt
# Mac端依赖包
redis==4.5.4
clickhouse-driver==0.2.6
PyYAML==6.0
Flask==2.3.2
Flask-CORS==4.0.0
pandas==2.0.3
